Ajout PermissionChecker from [Categorie => CompteRendu]

+++ src/Controller/CategorieController.php
@@ -5,7 +5,9 @@ namespace App\Controller;
 use App\Entity\Categorie;
 use App\Form\CategorieType;
 use App\Repository\CategorieRepository;
+use App\Services\Personne\PermissionChecker;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
@@ -15,14 +17,31 @@ use Symfony\Component\Routing\Annotation\Route;
  */
 class CategorieController extends AbstractController
 {
+    private $permissionChecker;
+
+    public function __construct(PermissionChecker $permissionChecker)
+    {
+        $this->permissionChecker = $permissionChecker;
+    }
+
     /**
      * @Route("/", name="categorie_index", methods={"GET"})
      */
     public function index(CategorieRepository $categorieRepository): Response
     {
-        return $this->render('categorie/index.html.twig', [
-            'categories' => $categorieRepository->findAll(),
-        ]);
+        if(!$this->isGranted('ROLE_USER')){
+            $this->addFlash('fail','Veuillez vous connectez pour acceder à cette page.');
+            return $this->redirectToRoute('app_login');
+        }else {
+            if (!$this->permissionChecker->isUserGranted(["GET_OTHER_CATEGORIE"])) {
+                $this->addFlash('fail', 'Vous ne possedez pas les permissions necessaires.');
+                return new RedirectResponse("/");
+            }else {
+                return $this->render('categorie/index.html.twig', [
+                    'categories' => $categorieRepository->findAll(),
+                ]);
+            }
+        }
     }
 
     /**
@@ -30,22 +49,32 @@ class CategorieController extends AbstractController
      */
     public function new(Request $request): Response
     {
-        $categorie = new Categorie();
-        $form = $this->createForm(CategorieType::class, $categorie);
-        $form->handleRequest($request);
+        if(!$this->isGranted('ROLE_USER')){
+            $this->addFlash('fail','Veuillez vous connectez pour acceder à cette page.');
+            return $this->redirectToRoute('app_login');
+        }else {
+            if (!$this->permissionChecker->isUserGranted(["POST_CATEGORIE"])) {
+                $this->addFlash('fail', 'Vous ne possedez pas les permissions necessaires.');
+                return new RedirectResponse("/");
+            } else {
+                $categorie = new Categorie();
+                $form = $this->createForm(CategorieType::class, $categorie);
+                $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->persist($categorie);
-            $entityManager->flush();
+                if ($form->isSubmitted() && $form->isValid()) {
+                    $entityManager = $this->getDoctrine()->getManager();
+                    $entityManager->persist($categorie);
+                    $entityManager->flush();
 
-            return $this->redirectToRoute('categorie_index');
+                    return $this->redirectToRoute('categorie_index');
+                }
+
+                return $this->render('categorie/new.html.twig', [
+                    'categorie' => $categorie,
+                    'form' => $form->createView(),
+                ]);
+            }
         }
-
-        return $this->render('categorie/new.html.twig', [
-            'categorie' => $categorie,
-            'form' => $form->createView(),
-        ]);
     }
 
     /**
@@ -53,9 +82,19 @@ class CategorieController extends AbstractController
      */
     public function show(Categorie $categorie): Response
     {
-        return $this->render('categorie/show.html.twig', [
-            'categorie' => $categorie,
-        ]);
+        if(!$this->isGranted('ROLE_USER')){
+            $this->addFlash('fail','Veuillez vous connectez pour acceder à cette page.');
+            return $this->redirectToRoute('app_login');
+        }else {
+            if (!$this->permissionChecker->isUserGranted(["GET_OTHER_CATEGORIE"])) {
+                $this->addFlash('fail', 'Vous ne possedez pas les permissions necessaires.');
+                return new RedirectResponse("/");
+            } else {
+                return $this->render('categorie/show.html.twig', [
+                    'categorie' => $categorie,
+                ]);
+            }
+        }
     }
 
     /**
@@ -63,19 +102,29 @@ class CategorieController extends AbstractController
      */
     public function edit(Request $request, Categorie $categorie): Response
     {
-        $form = $this->createForm(CategorieType::class, $categorie);
-        $form->handleRequest($request);
+        if(!$this->isGranted('ROLE_USER')){
+            $this->addFlash('fail','Veuillez vous connectez pour acceder à cette page.');
+            return $this->redirectToRoute('app_login');
+        }else {
+            if (!$this->permissionChecker->isUserGranted(["UPDATE_OTHER_CATEGORIE"])) {
+                $this->addFlash('fail', 'Vous ne possedez pas les permissions necessaires.');
+                return new RedirectResponse("/");
+            } else {
+                $form = $this->createForm(CategorieType::class, $categorie);
+                $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $this->getDoctrine()->getManager()->flush();
+                if ($form->isSubmitted() && $form->isValid()) {
+                    $this->getDoctrine()->getManager()->flush();
 
-            return $this->redirectToRoute('categorie_index');
+                    return $this->redirectToRoute('categorie_index');
+                }
+
+                return $this->render('categorie/edit.html.twig', [
+                    'categorie' => $categorie,
+                    'form' => $form->createView(),
+                ]);
+            }
         }
-
-        return $this->render('categorie/edit.html.twig', [
-            'categorie' => $categorie,
-            'form' => $form->createView(),
-        ]);
     }
 
     /**
@@ -83,12 +132,21 @@ class CategorieController extends AbstractController
      */
     public function delete(Request $request, Categorie $categorie): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$categorie->getId(), $request->request->get('_token'))) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->remove($categorie);
-            $entityManager->flush();
+        if (!$this->isGranted('ROLE_USER')) {
+            $this->addFlash('fail', 'Veuillez vous connectez pour acceder à cette page.');
+            return $this->redirectToRoute('app_login');
+        }else {
+            if (!$this->permissionChecker->isUserGranted(["DELETE_OTHER_CATEGORIE"])) {
+                $this->addFlash('fail', 'Vous ne possedez pas les permissions necessaires.');
+                return new RedirectResponse("/");
+            }else {
+                if ($this->isCsrfTokenValid('delete' . $categorie->getId(), $request->request->get('_token'))) {
+                    $entityManager = $this->getDoctrine()->getManager();
+                    $entityManager->remove($categorie);
+                    $entityManager->flush();
+                }
+                return $this->redirectToRoute('categorie_index');
+            }
         }
-
-        return $this->redirectToRoute('categorie_index');
     }
 }
