+++ src/Controller/CompteRenduController.php
@@ -14,13 +14,16 @@ use App\Repository\StatutRepository;
 use App\Services\Commune\CommuneService;
 use App\Services\CompteRendu\CompteRenduService;
 use App\Services\CompteRendu\DocumentService;
+use App\Services\Personne\PermissionChecker;
 use App\Services\UploadDocumentService;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpFoundation\Session\SessionInterface;
 use Symfony\Component\Routing\Annotation\Route;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
+use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
 
 
 /**
@@ -30,10 +33,19 @@ use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 class CompteRenduController extends AbstractController
 {
     private $problemeRepository;
+    private $compteRenduRepository;
+    private $personne;
+    /**
+     * @var PermissionChecker
+     */
+    private $permissionChecker;
 
-    public function __construct(ProblemeRepository $problemeRepository)
+    public function __construct(TokenStorageInterface $tokenStorageInterface, ProblemeRepository $problemeRepository, PermissionChecker $permissionChecker, CompteRenduRepository $compteRenduRepository)
     {
         $this->problemeRepository = $problemeRepository;
+        $this->permissionChecker = $permissionChecker;
+        $this->compteRenduRepository = $compteRenduRepository;
+        $this->personne = $tokenStorageInterface->getToken()->getUser();
     }
 
     /**
@@ -41,9 +53,19 @@ class CompteRenduController extends AbstractController
      */
     public function index(CompteRenduRepository $compteRenduRepository): Response
     {
-        return $this->render('compte_rendu/index.html.twig', [
-            'compte_rendus' => $compteRenduRepository->findAll(),
-        ]);
+        if(!$this->isGranted('ROLE_USER')){
+            $this->addFlash('fail','Veuillez vous connectez pour acceder à cette page.');
+            return $this->redirectToRoute('app_login');
+        }else {
+            if (!$this->permissionChecker->isUserGranted(["GET_OTHER_COMPTE_RENDU"])) {
+                $this->addFlash('fail', 'Vous ne possedez pas les permissions necessaires.');
+                return new RedirectResponse("/");
+            } else {
+                return $this->render('compte_rendu/index.html.twig', [
+                    'compte_rendus' => $compteRenduRepository->findAll(),
+                ]);
+            }
+        }
     }
 
     /**
@@ -51,21 +73,28 @@ class CompteRenduController extends AbstractController
      */
     public function nouveau(Request $request, SessionInterface $session): Response
     {
+        if(!$this->isGranted('ROLE_USER')){
+            $this->addFlash('fail','Veuillez vous connectez pour acceder à cette page.');
+            return $this->redirectToRoute('app_login');
+        }else {
+            if (!$this->permissionChecker->isUserGranted(["POST_COMPTE_RENDU"])) {
+                $this->addFlash('fail', 'Vous ne possedez pas les permissions necessaires.');
+                return new RedirectResponse("/");
+            } else {
+                $form = $this->createForm(ChoiceProblemType::class);
+                $form->handleRequest($request);
+                if ($form->isSubmitted() && $form->isValid()) {
+                    $problemeId = (int)$request->request->all()['choice_problem']["Probleme"];
+                    $probleme = $this->problemeRepository->find($problemeId);
+                    $session->set('Probleme', $probleme);
 
-        $form = $this->createForm(ChoiceProblemType::class);
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $problemeId = (int)$request->request->all()['choice_problem']["Probleme"];
-            $probleme = $this->problemeRepository->find($problemeId);
-            $session->set('Probleme', $probleme);
-
-            return $this->redirectToRoute('compte_rendu_new');
+                    return $this->redirectToRoute('compte_rendu_new');
+                }
+                return $this->render('compte_rendu/choiceProblemNew.html.twig', [
+                    'form' => $form->createView(),
+                ]);
+            }
         }
-
-        return $this->render('compte_rendu/choiceProblemNew.html.twig', [
-            'form' => $form->createView(),
-        ]);
     }
 
     /**
@@ -79,31 +108,41 @@ class CompteRenduController extends AbstractController
         CompteRenduService $compteRenduService
     ): Response
     {
-        $probleme = $session->get('Probleme');
-        if($probleme == null) {
-            $this->addFlash('fail','Aucun problème n\'a été selectionné');
-            return $this->redirectToRoute('compte_rendu_nouveau');
+        if(!$this->isGranted('ROLE_USER')){
+            $this->addFlash('fail','Veuillez vous connectez pour acceder à cette page.');
+            return $this->redirectToRoute('app_login');
+        }else {
+            if (!$this->permissionChecker->isUserGranted(["POST_COMPTE_RENDU"])) {
+                $this->addFlash('fail', 'Vous ne possedez pas les permissions necessaires.');
+                return new RedirectResponse("/");
+            } else {
+                $probleme = $session->get('Probleme');
+                if ($probleme == null) {
+                    $this->addFlash('fail', 'Aucun problème n\'a été selectionné');
+                    return $this->redirectToRoute('compte_rendu_nouveau');
+                }
+                $compteRendu = new CompteRendu();
+                $pb = $problemeRepository->findOneBy(['id' => $probleme->getId()]);
+                $compteRendu->setProbleme($pb);
+                $form = $this->createForm(CompteRenduType::class, $compteRendu);
+                $form->handleRequest($request);
+
+                if ($form->isSubmitted() && $form->isValid()) {
+                    $document = $form->get('urlDocument')->getData();
+                    $compteRenduService->PersistCompteRendu($compteRendu, $pb, $document);
+                    $entityManager = $this->getDoctrine()->getManager();
+                    $entityManager->persist($compteRendu);
+                    $entityManager->flush();
+
+                    return $this->redirectToRoute('compte_rendu_index');
+                }
+
+                return $this->render('compte_rendu/new.html.twig', [
+                    'compte_rendu' => $compteRendu,
+                    'form' => $form->createView(),
+                ]);
+            }
         }
-        $compteRendu = new CompteRendu();
-        $pb = $problemeRepository->findOneBy(['id'=> $probleme->getId()]);
-        $compteRendu->setProbleme($pb);
-        $form = $this->createForm(CompteRenduType::class, $compteRendu);
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $document = $form->get('urlDocument')->getData();
-            $compteRenduService->PersistCompteRendu($compteRendu,$pb,$document);
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->persist($compteRendu);
-            $entityManager->flush();
-
-            return $this->redirectToRoute('compte_rendu_index');
-        }
-
-        return $this->render('compte_rendu/new.html.twig', [
-            'compte_rendu' => $compteRendu,
-            'form' => $form->createView(),
-        ]);
     }
 
     /**
@@ -111,9 +150,25 @@ class CompteRenduController extends AbstractController
      */
     public function show(CompteRendu $compteRendu): Response
     {
-        return $this->render('compte_rendu/show.html.twig', [
-            'compte_rendu' => $compteRendu,
-        ]);
+        $request = $this->compteRenduRepository->getOneCompteRenduByTechnicien($compteRendu, $this->personne);
+        if(!$this->isGranted('ROLE_USER')){
+            $this->addFlash('fail','Veuillez vous connectez pour acceder à cette page.');
+            return $this->redirectToRoute('app_login');
+        }else {
+            if(!$request) {
+                $this->addFlash('fail', 'Ce compte rendu ne vous appartient pas.');
+                return $this->redirectToRoute("home_index");
+            }else {
+                if (!$this->permissionChecker->isUserGranted(["GET_OTHER_COMPTE_RENDU"])) {
+                    $this->addFlash('fail', 'Vous ne possedez pas les permissions necessaires.');
+                    return new RedirectResponse("/");
+                }else {
+                    return $this->render('compte_rendu/show.html.twig', [
+                        'compte_rendu' => $compteRendu,
+                    ]);
+                }
+            }
+        }
     }
 
     /**
@@ -121,19 +176,29 @@ class CompteRenduController extends AbstractController
      */
     public function edit(Request $request, CompteRendu $compteRendu): Response
     {
-        $form = $this->createForm(CompteRenduType::class, $compteRendu);
-        $form->handleRequest($request);
+        if(!$this->isGranted('ROLE_USER')){
+            $this->addFlash('fail','Veuillez vous connectez pour acceder à cette page.');
+            return $this->redirectToRoute('app_login');
+        }else {
+            if (!$this->permissionChecker->isUserGranted(["UPDATE_COMPTE_RENDU"])) {
+                $this->addFlash('fail', 'Vous ne possedez pas les permissions necessaires.');
+                return new RedirectResponse("/");
+            } else {
+                $form = $this->createForm(CompteRenduType::class, $compteRendu);
+                $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $this->getDoctrine()->getManager()->flush();
+                if ($form->isSubmitted() && $form->isValid()) {
+                    $this->getDoctrine()->getManager()->flush();
 
-            return $this->redirectToRoute('compte_rendu_index');
+                    return $this->redirectToRoute('compte_rendu_index');
+                }
+
+                return $this->render('compte_rendu/edit.html.twig', [
+                    'compte_rendu' => $compteRendu,
+                    'form' => $form->createView(),
+                ]);
+            }
         }
-
-        return $this->render('compte_rendu/edit.html.twig', [
-            'compte_rendu' => $compteRendu,
-            'form' => $form->createView(),
-        ]);
     }
 
     /**
@@ -141,13 +206,23 @@ class CompteRenduController extends AbstractController
      */
     public function delete(Request $request, CompteRendu $compteRendu): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$compteRendu->getId(), $request->request->get('_token'))) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->remove($compteRendu);
-            $entityManager->flush();
-        }
+        if (!$this->isGranted('ROLE_USER')) {
+            $this->addFlash('fail', 'Veuillez vous connectez pour acceder à cette page.');
+            return $this->redirectToRoute('app_login');
+        } else {
+            if (!$this->permissionChecker->isUserGranted(["DELETE_COMPTE_RENDU"])) {
+                $this->addFlash('fail', 'Vous ne possedez pas les permissions necessaires.');
+                return new RedirectResponse("/");
+            } else {
+                if ($this->isCsrfTokenValid('delete' . $compteRendu->getId(), $request->request->get('_token'))) {
+                    $entityManager = $this->getDoctrine()->getManager();
+                    $entityManager->remove($compteRendu);
+                    $entityManager->flush();
+                }
 
-        return $this->redirectToRoute('compte_rendu_index');
+                return $this->redirectToRoute('compte_rendu_index');
+            }
+        }
     }
 
 }
